Seperate sleep on power button test
