+++ tests/integration/modules/mac_power.py
@@ -144,56 +144,6 @@ class MacPowerModuleTest(integration.ModuleCase):
 @skipIf(not salt.utils.is_darwin()
         or not salt.utils.which('systemsetup')
         or salt.utils.get_uid(salt.utils.get_user()) != 0, 'Test requirements not met')
-class MacPowerModuleTestSleepOnPowerButton(integration.ModuleCase):
-    '''
-    Test power.get_sleep_on_power_button
-    Test power.set_sleep_on_power_button
-    '''
-    SLEEP_ON_BUTTON = None
-
-    def setup(self):
-        '''
-        Check if function is available
-        Get existing value
-        '''
-        # Is the function available
-        ret = self.run_function('power.get_sleep_on_power_button')
-        if isinstance(ret, bool):
-            self.SLEEP_ON_BUTTON = self.run_function(
-                'power.get_sleep_on_power_button')
-
-    def teardown(self):
-        '''
-        Reset to original value
-        '''
-        if self.SLEEP_ON_BUTTON is not None:
-            self.run_function(
-                'power.set_sleep_on_power_button', [self.SLEEP_ON_BUTTON])
-
-    def test_sleep_on_power_button(self):
-        '''
-        Test power.get_sleep_on_power_button
-        Test power.set_sleep_on_power_button
-        '''
-        # If available on this system, test it
-        if self.SLEEP_ON_BUTTON is None:
-            # Check for not available
-            ret = self.run_function('power.get_sleep_on_power_button')
-            self.assertIn('Error', ret)
-        else:
-            self.assertTrue(
-                self.run_function('power.set_sleep_on_power_button', ['on']))
-            self.assertTrue(
-                self.run_function('power.get_sleep_on_power_button'))
-            self.assertTrue(
-                self.run_function('power.set_sleep_on_power_button', ['off']))
-            self.assertFalse(
-                self.run_function('power.get_sleep_on_power_button'))
-
-
-@skipIf(not salt.utils.is_darwin()
-        or not salt.utils.which('systemsetup')
-        or salt.utils.get_uid(salt.utils.get_user()) != 0, 'Test requirements not met')
 class MacPowerModuleTestRestartPowerFailure(integration.ModuleCase):
     '''
     Test power.get_restart_power_failure
@@ -335,7 +285,6 @@ class MacPowerModuleTestWakeOnModem(integration.ModuleCase):
 if __name__ == '__main__':
     from integration import run_tests
     run_tests(MacPowerModuleTest,
-              MacPowerModuleTestSleepOnPowerButton,
               MacPowerModuleTestRestartPowerFailure,
               MacPowerModuleTestWakeOnNet,
               MacPowerModuleTestWakeOnModem)
